docker_gcp/requirements.txt
google-cloud-storage==1.25.0
pandas==0.25.3
